backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Decomp.me backend Django application"
authors = [
    {name = "Decomp.me team"}
]
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "django>=5.1.3",
    "django-cors-headers>=4.6.0",
    "django-environ>=0.11.2",
    "django-filter>=24.3",
    "djangorestframework>=3.15.2",
    "psycopg2-binary>=2.9.10",
    "jwt>=1.3.1",
    "PyGithub>=2.5.0",
    "drf-extensions>=0.7.1",
    "tzdata==2024.2",
    "Pillow>=11.0",
    "html-json-forms>=1.1.1",
    "django-resized>=1.0.3",
    "django-cleanup>=9.0.0",
    "m2c @ git+https://github.com/matt-kempster/m2c.git",
    "asm-differ @ git+https://github.com/simonlindholm/asm-differ.git",
    "sentry-sdk>=2.19.0",
    "django-session-timeout>=0.1.0",
    "requests>=2.32.3",
    "setuptools>=75.6.0",
    "debugpy>=1.8.13",
    "gunicorn>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.13.1",
    "usort>=1.0.8.post1",
    "django-stubs-ext>=5.1.1",
    "django-stubs>=5.1.1",
    "djangorestframework-stubs>=3.15.1",
    "types-requests>=2.32.0.20241016",
    "parameterized>=0.9.0",
    "mypy>=1.13.0",
    "responses==0.25.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["coreapp", "decompme"]

[tool.ruff]
extend-exclude = ["local_files"]
line-length = 88

[tool.ruff.format]
docstring-code-format = true

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
check_untyped_defs = true
disallow_any_generics = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
implicit_reexport = false
strict_equality = true
no_implicit_optional = true
warn_unused_configs = true
warn_unreachable = true
warn_no_return = true
namespace_packages = true
disallow_untyped_defs = true

files = [
    "coreapp/**/*.py",
    "decompme/**/*.py"
]
exclude = ["manage.py"]

plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main"
]

[tool.django-stubs]
django_settings_module = "decompme.settings"
